Upgrade go version and go deps (#633)

+++ swagger_gen/models/tag.go
@@ -89,7 +89,7 @@ func (m *Tag) ContextValidate(ctx context.Context, formats strfmt.Registry) erro
 
 func (m *Tag) contextValidateID(ctx context.Context, formats strfmt.Registry) error {
 
-	if err := validate.ReadOnly(ctx, "id", "body", int64(m.ID)); err != nil {
+	if err := validate.ReadOnly(ctx, "id", "body", m.ID); err != nil {
 		return err
 	}
